+++ TShockAPI/Permissions.cs
@@ -128,6 +128,9 @@ namespace TShockAPI
 		[Description("User can download updates to plugins that are currently running.")]
 		public static readonly string updateplugins = "tshock.cfg.updateplugins";
 
+		[Description("User can create reference files of Terraria IDs and the permission matrix in the server folder.")]
+		public static readonly string createdumps = "tshock.cfg.createdumps";
+
 		// tshock.ignore nodes
 
 		[Description("Prevents you from being reverted by kill tile abuse detection.")]